requirements.txt
# Core dependencies
numpy>=1.22.0
scipy>=1.8.0
pandas>=1.4.0
matplotlib>=3.5.0

# EEG processing and ML
scikit-learn>=1.0.0
mne>=1.0.0
pywavelets>=1.3.0

# Vector Database
supabase>=1.0.3
postgrest>=0.10.6
httpx>=0.23.0
realtime>=0.0.4

# Gemini Integration (replacing OpenAI)
google-generativeai>=0.3.0
requests>=2.28.0

# LangChain (Optional)
langchain>=0.0.267
langchain-google-genai>=0.0.3  # Replacing langchain-openai

# Testing and utilities
pytest>=7.0.0
pytest-cov>=2.12.1
tqdm>=4.64.0
joblib>=1.1.0
pyyaml>=6.0

# Machine Learning
torch>=1.10.0
transformers>=4.15.0
tensorflow>=2.10.0  # Required for autoencoder tokenizer
keras>=2.10.0       # Higher-level API for neural networks

# API and Web
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0 